+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-- First person hand rendering now matches vanilla
-- Fixed an issue where skins would not work in multiplayer when loaded from a resource pack
+- Added "Auto Selected" skin to the default skin pack. Selecting this skin will reset your skin to its default appearance. (Legacy4J only)
+- Removed Reset Skin keybind, as it is no longer needed. (Legacy4J only)
 
 **Full Changelog**: https://github.com/BrandonItaly/Bedrock-Skins/commits/master